requirements.txt
# Core Dependencies
tensorflow>=2.3.0,<3.0.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0

# Data Processing
scipy>=1.5.0
pandas>=1.1.0

# Configuration
PyYAML>=5.3.0

# Image Processing
opencv-python>=4.5.0

# Optional: For YAD2K utilities (if using original implementation)
h5py>=2.10.0
keras>=2.4.0
